+++ tests/testthat/test-clone.R
@@ -273,7 +273,6 @@ test_that("Cloning active bindings", {
 })
 
 
-
 test_that("Cloning active binding in superclass", {
   AC <- R6Class("AC",
     public = list(
@@ -315,6 +314,44 @@ test_that("Cloning active binding in superclass", {
   expect_identical(b1$x2, 48)
 })
 
+
+test_that("Cloning active binding in two levels of inheritance", {
+  # For issue #119
+  A <- R6Class("A",
+    public = list(
+      methodA = function() "A"
+    ),
+    active = list(
+      x = function() "x"
+    )
+  )
+
+  B <- R6Class("B",
+    inherit = A,
+    public = list(
+      methodB = function() {
+        super$methodA()
+      }
+    )
+  )
+
+  C <- R6Class("C",
+    inherit = B,
+    public = list(
+      methodC = function() {
+        super$methodB()
+      }
+    )
+  )
+
+  C1 <- C$new()
+  C2 <- C1$clone()
+  expect_identical(C2$methodC(), "A")
+  expect_identical(
+    C1$.__enclos_env__$super$.__enclos_env__,
+    environment(C1$.__enclos_env__$super$methodB)
+  )
+})
 
 
 test_that("Lock state", {